pyproject.toml
[project]
name = "exporter"
version = "0.1.0"
description = "Prometheus-Exporter and tiny emulator, for mmp14 baytech PDUs"
readme = "README.md"
requires-python = ">=3.13"
authors = [
  {name = "Janik Haag", email = "janik@aq0.de"},
]
dependencies = [
    "prometheus-client>=0.22.1",
    "pyserial>=3.5",
]

[dependency-groups]
dev = [ "pytest>=8.4.2" ]

[project.scripts]
mmp-exporter = "src.exporter.main:main"
mmp-emulator = "src.emulator.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["./src"]
